I listen to user feedback
Main problem was the requirement of c++ redist so you couldn't crash efficiently on freshly installed vms.

Also adjusted timings for the ear destruction audio and got rid of the console window since it was too easy to have second thoughts for some users.

give me a phd for writing my thesis thanks